Refactored window out of the app
